refactor: I put together the providers.
