+++ frontend/src/index.tsx
@@ -1,19 +1,18 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
-import { ThemeProvider } from 'styled-components'
-import { theme } from 'styles/theme'
+import { Providers } from 'providers/Providers'
 import { GlobalStyle } from 'styles/globalStyle'
 import { ToastContainer } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
 
 ReactDOM.render(
   <React.StrictMode>
-    <ThemeProvider theme={theme}>
+    <Providers>
       <GlobalStyle />
       <ToastContainer />
       <App />
-    </ThemeProvider>
+    </Providers>
   </React.StrictMode>,
   document.getElementById('root'),
 )
