+++ frontend/src/pages/projectList/projectDetail/ProjectDetail.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useEffect, useState } from 'react'
 import { Navigate, NavLink, useParams } from 'react-router-dom'
-import { useAuthContext } from 'context/AuthProvider'
+import { useAuthContext } from 'providers/AuthProvider'
 import { useGetCurrentUserLazyQuery } from './getUser.gen'
 import {
   useGetProjectLazyQuery,
